+++ scripts/build.sh
@@ -2,11 +2,11 @@
 # vim: ft=sh ts=4 sw=4 sts=4 noet
 set -euf
 
-# Alias docker-compose with params we need for the test environment.
-alias doco='docker-compose --project-name dab -f tests/docker-compose.yml'
+export COMPOSE_PROJECT_NAME='dab'
+export COMPOSE_FILE="tests/docker-compose.yml:tests/docker-compose.${TEST_DOCKER:-local}.yml"
 
 # Pull the last dab image for caching.
-doco pull dab || true
+docker-compose pull dab || true
 
 # Build just the dab image and clean up.
-doco build --force-rm --pull dab
+docker-compose build --force-rm --pull dab